Remove some errors
