+++ src/components/SlideList.jsx
@@ -1,6 +1,13 @@
 
 const SlideView = ({selectedSlide, slides}) => {
-  // return JSX for slide view
+   return (
+   <div className="col-span-2 border border-gray-300 p-4 rounded-lg"> {/* middle column */}
+    <div className="flex justify-center w-full">
+      <div className="w-3/4">
+        {selectedSlide !== null && slides[selectedSlide] ? displaySlide(slides[selectedSlide]) : null }
+      </div>
+    </div>
+  </div>)
 }
 
 export default SlideView;
